Fix env
